test: move resources-e2e test to the kafka lane (always-on advertising needs a broker)
`test_resources_reach_agent_run_and_agent_tool` hosts a function tool node, which
now always advertises on `calf.capabilities`. At `worker.start()` the worker stands
up a real `GroupedKafkaTableWriter` for the control plane, and that writer opens its
own aiokafka producer — which `TestKafkaBroker` does NOT patch. So the offline test
needed a reachable broker and failed in CI's offline lane (it only passed locally
because a Redpanda broker happened to be listening on :9092).

Move it to `tests/integration/test_lifecycle_resources_kafka.py` (`pytestmark=kafka`,
`ALLOW_MODEL_REQUESTS=True`, the `kafka_bootstrap` fixture -> `Client.connect(bootstrap)`).
The agent->tool round trip still runs over `TestKafkaBroker`; the real broker is needed
only so the control-plane writer can start. Update the offline file's now-stale module
docstring (it no longer covers the Agent-run / agent_tool surfaces) and drop the imports
that moved with the test.

Verified: offline lane 3181 passed; the moved test passes against a real broker via
testcontainers (CI parity).
